+++ src/redux/modules/entities/selectors.js
@@ -59,7 +59,7 @@ export const userHasCreatePermission = state =>
     getCurrentEntityStore(state).get('createPermission')
   );
 
-const hasPermission = (userPermissions, permissionsNeeded) => {
+export const hasPermission = (userPermissions, permissionsNeeded) => {
   if (permissionsNeeded !== undefined) {
     // Return true if they have at least one of the permissions
     return permissionsNeeded.some(permissionNeeded =>
@@ -78,9 +78,7 @@ export const isCreating = state =>
 
 export const isUpdating = state => getCurrentEntityStore(state).get('updating');
 
-export const isSaving = state =>
-  isCreating(state) ||
-  isUpdating(state)
+export const isSaving = state => isCreating(state) || isUpdating(state);
 
 export const getCurrentSubEntity = state =>
   getCurrentEntityStore(state).get('subEntity');